Use GumV8InterceptorIgnoreScope and simplify the code
